api/requirements.txt
# ── Web framework ─────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── HTTP client ───────────────────────────────────────────────
httpx==0.27.0

# ── Redis ─────────────────────────────────────────────────────
redis[asyncio]==5.0.4

# ── Data validation ───────────────────────────────────────────
pydantic==1.10.17

# ── DNS checks (domain health) ────────────────────────────────
dnspython==2.6.1

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.0.1
